requirements.txt
scipy
scikit-learn
numpy
opencv-python
Pillow
torch
torchvision
flask